move MoleculeLJ to md subfolder
